sampleapi/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlmodel
aiosqlite
greenlet
passlib[bcrypt]
python-jose
pytest
httpx